==> scRNA-seq/01. LLM label assignment.ipynb <==
# Annotate Seurat's "FindMarkers" upregulated genes using GPT4 LLM

function calling allows us to use parasable json structured data

### Add openai's python module

!pip install -q openai

import os
import json
import pandas as pd
from getpass import getpass

import openai

### Get the user's API key

# Prompt for secure key input
api_key = getpass('Enter your OPENAI API key: ')

# Set the key as an environment variable

os.environ['OPENAI_API_KEY'] = api_key

### Read the markers table

in this example we have saved rownames on export from R

markers = pd.read_csv('/sbgenomics/project-files/scRNAseq_processing/seurat_scRNA_filtered_downsampled.harmonyJLW.markers.csv',index_col=0)

markers.head()

### Set parameters for genes that will inform clusters

n_genes = 20
orderby_columns = ['avg_log2FC','p_val']
orderby_ascending = [False,True]
max_p_value_adj = 0.05
min_avg_log2FC = 0.25
drop_gene_strings = ['chr']

ordered_markers = markers.sort_values(orderby_columns,ascending=orderby_ascending).reset_index(drop=True)
for ds in drop_gene_strings:
    ordered_markers = ordered_markers.loc[~ordered_markers['gene'].str.contains(ds),:]
ordered_markers = ordered_markers.loc[ordered_markers['p_val_adj'] < max_p_value_adj,:]
print("Number of clusters:")
print(len(ordered_markers['cluster'].unique()))
cluster_obj = {}
for i in sorted(ordered_markers['cluster'].unique()):
    sub = ordered_markers.loc[ordered_markers['cluster']==i,:]
    glist = list(sub.head(n=n_genes)['gene'])
    cluster_obj[i] = glist

### Set up our GPT function

GPT_MODEL = "gpt-4"


cell_types_allowed = ["TUFT", "ENDOCRINE", "B/PLASMA", "GRANULOCYTE", "EPITHELIAL", "ACINAR", "MYELOID", "T/NK", "PERICYTE", "FIBROBLASTS", "SMOOTH MUSCLE", "DENDRITIC", "ENDOTHELIAL", "MAST CELL", "RED BLOOD CELL"]
context = """
Given this sell type rough categories and a few example genes:
'POU2F3','VAV1', #TUFT
                                           'INSR','CHGA',"TTR", #ENDOCRINE
                                           'CD79A','MS4A1','IGJ','IGLL5', #B/PLASMA
                                           "G0S2", 'CXCR2',"CXCR1", #GRANULOCYTE
                                           'KRT18','KRT8','TFF1','KRT19','SPINK1', #EPITHELIAL
                                           "PRSS1",'AMY2A','CTRB2','REG1A',#ACINAR
                                           'LYZ','APOE', 'HLA-DRA', 'C1QA','CD14', "CD68", #MYELOID
                                           'CD2','CD3D','NKG7','NCAM1','CD8A','CD4','FOXP3', #T AND NK
                                           'IGFBP7','ACTA2','RGS5', #PERICYTE
                                           'COL1A1','DCN','LUM','CDH11', #FIBROBLASTS
                                           'TAGLN','MYL9', #SMOOTH MUSCLE
                                           'IRF7','GZMB','CCL22','LAMP3',"HLA-DQA1", #DENDRITIC
                                           'PLVAP','VWF', #ENDOTHELIAL
                                           'TPSAB1','CPA3', #MAST CELL
                                           'HBB','HBA2' #RED BLOOD CELL

We want to classify some clusters as being either
TUFT, ENDOCRINE, B/PLASMA, GRANULOCYTE, EPITHELIAL, ACINAR, MYELOID, T/NK, PERICYTE, FIBROBLASTS, SMOOTH MUSCLE, DENDRITIC, ENDOTHELIAL, MAST CELL, or RED BLOOD CELL (just one that is the best match)
""".strip()

tools = [
  {
    "type": "function",
    "function": {
      "name": "cluster_assignment",
      "description": "Relay the decision on which cell type the genes provided best describe.",
      "parameters": {
        "type": "object",
        "properties": {
          "evidence": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "gene_name": {
                  "type": "string",
                  "description": "A gene that is informative for selecting the cell type"
                },
                "reason_to_use": {
                  "type": "string",
                  "description": "The reason this gene is informative for selecting the cell type"
                }
              },
              "required": [
                "gene_name",
                "reason_to_use"
              ]
            },
            "description": "A list of genes and their evidence that are helpful for determining cell type, do NOT list genes that are not helpful int he final cell type assignment."
          },
          "cell_type_assignment": {
            "type": "string",
            "enum": cell_types_allowed + ([] if "UNKNOWN" in cell_types_allowed else ["UNKNOWN"]),
            "description": "The most likely cell type if known."
          }
        },
        "required": [
          "evidence",
          "cell_type_assignment"
        ]
      }
    }
  }
]



### Iterate through and label using GPT

def make_messages(glist):
    messages = []
    messages.append({"role":"system","content":"""
Your job is to determine which upregulated genes 
from a list give best evidence for a cluster give evidence of that cluster representing a specific cell type from among a specific list of cell types. 
Do not guess and take your time to reason the evidence for which genes work best.  Choose type 'UNKONWN' if its mixed or unkonwn.""".strip()
})
    messages.append({"role":"user", "content":"""
For each cluster I present you will have a list of some highly upregulated genes for each cluster,
first say which genes are most informative for a cell type and why, then choose one most appropriate cell type for it:
                """.strip()+"\n"+" ,".join(glist)})
    return messages

print(json.dumps(tools,indent=2))

client = openai.OpenAI()
cluster_assignment = {}
for i, glist in cluster_obj.items():
    use_messages = make_messages(glist)
    response = client.chat.completions.create(
        model=GPT_MODEL,
        messages=make_messages(glist),
        tools = tools,
       tool_choice={"type": "function", "function": {"name": "cluster_assignment"}}
    )
    try:
        assignment = json.loads(json.loads(response.model_dump_json())['choices'][0]['message']['tool_calls'][0]['function']['arguments'])
    except:
        assignment = {'evidence':[],'cell_type_assignment':'UNKNOWN'}
    print("------------")
    print(f"CLUSTER {i}")
    print(json.dumps(assignment,indent=2))
    cluster_assignment[str(i)] = assignment #for string for json compatiability

with open('/sbgenomics/output-files/cell_type_assignment2.gpt4.json','w') as of:
    of.write(json.dumps(cluster_assignment,indent=2))
